python/requirements.txt
Flask==2.2.2
crypto==1.4.1
pycryptodome==3.16.0
gunicorn==20.1.0
PyJWT==2.3.0
redis==3.5.3